Added views to create and query bonds
- Not including handling legal_name in this commit.
- Add the Bond model: assume that field values shouldn't be longer than
the given example. It's not clear from the exercise brief whether any
of the field values should be unique, so use the default primary key
 'id' field.
- Add the Bond serializer. I'm not sure what validation we might
want based on the brief. For example perhaps we should add a
validator method to check maturity date is in the future. Perhaps
explicit allowed currency values should be stored in a table. I'll
consider these validations out of scope for now.
- Add the Bond ViewSet: allow users to create or list bonds. For
security, be explicit about what query params can be used and ignore
invalid query parameters.
- Restrict the number of results to return in a single list for scalability.