Lookup the legal_name
- Lookup the legal name in the Gleif API when creating a Bond in the
DB. I've assumed that the legal name for a given lei won't
change and it's ok to persist in the DB. With that same assumption
it would be sensible to cache or persist known mappings of leis to
legal_names to prevent multiple duplicate lookups. I'll consider that
 out of scope for now at least.
- Catch errors which come from the remote API and log them. I haven't
configured the logger to write to disk, it seems out of scope for the
brief; but the principle is that such logs would be useful for
debugging purposes, but those details shouldn't be returned to the
user of our API, for security purposes.
- Add requests lib to requirements.